+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The public API of this library consists of the functions declared in file
 [h3api.h.in](./src/h3lib/include/h3api.h.in).
 
 ## [Unreleased]
+### Fixed
+- `compact` handles zero length input correctly.
 
 ## [3.6.0] - 2019-08-12
 ### Added